requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
ddgs==9.13.1
duckduckgo_search==8.1.1
fastapi==0.135.3
filelock==3.28.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.4
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
primp==1.2.2
protobuf==7.34.1
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pytesseract==0.3.13
pytest==9.0.3
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
sentencepiece==0.2.1
shellingham==1.5.4
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
tavily-python==0.7.23
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
playwright==1.56.0
trafilatura==2.0.0
